requirements.txt
tensorflow-gpu==2.5.2
pandas==1.3.5
matplotlib==3.5.3 
scikit-learn==1.0.2
gensim==4.2.0   
tqdm
#conda install pydot
